fix[JSONSerializer]: fix unserializing of JSON data
1. The json module does not support memoryview() objects to load().

>>> import shared_memory_dict
>>> s = shared_memory_dict.SharedMemoryDict('test', 1000, serializer=shared_memory_dict.serializers.JSONSerializer())
>>> s
Fail to load data: TypeError('the JSON object must be str, bytes or bytearray, not memoryview')
{}
>>> s['foo']  = 1
Fail to load data: TypeError('the JSON object must be str, bytes or bytearray, not memoryview')
>>> s
Fail to load data: TypeError('the JSON object must be str, bytes or bytearray, not memoryview')
{}

2. The memoryview() data initially contains only null bytes (with the
size of the data)

3. If the size decreases, the memory is not overwritten with null bytes.
Therefore invalid JSON data is part of the memory.
Always append a null byte to the dumped data and read only until a null
byte when loading data.

Fixes: d1a564d24ae474d0f1f1d5ff2e16bdf2047919a9